Add typings, .env file, divide programs into modules, and make small optimization/clean code changes
In this commit I added typings, created .env file to store environment variables for api, divided programs into modules to make codebase organized and clean and made various small optimizations and improvements to enhance performance and readability.
