+++ .env
@@ -1,0 +1,3 @@
+API_URL = "https://random-words-api-umber.vercel.app/word"
+X_RAPIDAPI_KEY = "f934e406b8msh16f9dee78137dfcp1fab91jsn560d9925327d"
+X_RAPIDAPI_HOST = "random-words-with-pronunciation.p.rapidapi.com"
